small fixes

+++ src/main/java/com/eins/learn/algo/cryptography/rsa/KeyGenerator.java
@@ -17,8 +17,8 @@ public class KeyGenerator {
         int p = getPrime(bitLength, 0);
         int q = getPrime(bitLength, p);
 
-        long n = p * q;
-        long fi = (p - 1) * (q - 1);
+        long n = (long) p * q;
+        long fi = (long) (p - 1) * (q - 1);
 
         int e = calculateE(fi, 3);
         long d = calculateD(fi, e);
@@ -28,8 +28,8 @@ public class KeyGenerator {
             if (e > 100) {
                 p = getPrime(bitLength, 0);
                 q = getPrime(bitLength, p);
-                n = p * q;
-                fi = (p - 1) * (q - 1);
+                n = (long) p * q;
+                fi = (long) (p - 1) * (q - 1);
             }
         }
         Key publicKey = new Key(n, e);
